+++ common/JackLibAPI.cpp
@@ -49,7 +49,7 @@ extern "C"
 #endif
 
 JackLibGlobals* JackLibGlobals::fGlobals = NULL;
-long JackLibGlobals::fClientCount = 0;
+int JackLibGlobals::fClientCount = 0;
 
 EXPORT jack_client_t* jack_client_open(const char* ext_client_name, jack_options_t options, jack_status_t* status, ...)
 {
