+++ common/JackShmMem.cpp
@@ -25,7 +25,7 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 namespace Jack
 {
 
-unsigned long JackShmMem::fSegmentNum = 0;
+unsigned int JackShmMem::fSegmentNum = 0;
 jack_shm_info_t JackShmMem::gInfo;
 size_t JackMem::gSize = 0;
